cs425/CS425-Airline-Flight-Booking-Application/requirements.txt
Click==7.0
Faker==2.0.3
Flask==1.1.1
Flask-Cors==3.0.8
itsdangerous==1.1.0
Jinja2==2.10.3
MarkupSafe==1.1.1
psycopg2==2.8.4
python-dateutil==2.8.1
python-dotenv==0.10.3
six==1.13.0
text-unidecode==1.3
Werkzeug==0.16.0
